tcl_SimpleUnitTestLibrary updated to V1.0.4.0,  tcl_UnitTestBaseClasses V1.0.0.0 created
TUnitTestBaseClass has been moved to its own library (tcl_UnitTestBaseClasses) so that tests can be defined alongside the classes to be tested in a library. To do this
a reference to tcl_UnitTestBaseClasses must be added to the library project (but not tcl_SimpleUnitTestLibrary).

A class (TTestSuite) has been added to hold a list of related tests - tests are now added to this (not the TestRunner) by magic in the FB_init method of the test  during initialisation.
Test Suites must then be added programmatically to the Test Runner. This allows tests to be added/removed from the TestRunner as a group.

A plc project may just use tcl_SimpleUnitTestLibrary to define tests in a PLC project and run them as before (tcl_UnitTestBaseClasses is not required for this).